Version 2 MC Model
Overhauled, and has 5 and 6 ping variants. Better error estimate for
pings. All pings (not exact data, but I believe I have it close enough),
or just the last ping. (What I earlier said was 5th, and seemed to be as
of March 15th when I started the model; now appears to be 6th from
timeline.) More details included in writeup and analysis.